Move equal to utils
